+++ src/main/java/com/bot/spider/services/InputProcess/ProcessEntrance.java
@@ -1,6 +1,6 @@
 package com.bot.spider.services.InputProcess;
 
-import com.bot.spider.dtos.telegram.TelegramUpdateDTO;
+import com.bot.spider.dtos.telegram.update.TelegramUpdateDTO;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
